+++ src/app/auth/forgot-password/forgot-password.component.ts
@@ -5,7 +5,7 @@ import { MatSnackBar } from "@angular/material";
 @Component({
   selector: "app-forgot-password",
   templateUrl: "./forgot-password.component.html",
-  styleUrls: ["./forgot-password.component.scss"]
+  styleUrls: ["./forgot-password.component.scss", "./../shared-auth.scss"]
 })
 export class ForgotPasswordComponent implements OnInit {
   emailAddress: string;