+++ src/app/auth/verify-email/verify-email.component.ts
@@ -5,7 +5,7 @@ import { MatSnackBar } from "@angular/material";
 @Component({
   selector: "app-verify-email",
   templateUrl: "./verify-email.component.html",
-  styleUrls: ["./verify-email.component.scss"]
+  styleUrls: ["./verify-email.component.scss", "./../shared-auth.scss"]
 })
 export class VerifyEmailComponent implements OnInit {
   constructor(public userService: UserService, public snackBar: MatSnackBar) {}
@@ -16,5 +16,9 @@ export class VerifyEmailComponent implements OnInit {
     this.snackBar.open("Verification email sent", "Close", {
       duration: 5000
     });
+  }
+
+  isSendingVerificationEmail(): boolean {
+    return this.userService.user && !this.userService.user.emailVerified;
   }
 }